weekender/requirements.txt
# LangGraph Concert Agent Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langsmith>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
